LPD-85800 Fold ScriptingValidator into ScriptingExecutor
# breaking

## What portal-kernel/src/com/liferay/portal/kernel/scripting/ScriptingValidator.java

The interface `ScriptingValidator` has been removed. Its
`validate(String)` method has been added to `ScriptingExecutor`.

## Why

The validator was a parallel SPI to `ScriptingExecutor` with the same
per-language registration pattern. Collapsing it onto `ScriptingExecutor`
removes the duplicate service tracker and keeps all language-specific
scripting behavior on a single interface.

----
