+++ portal-kernel/src/com/liferay/portal/kernel/scripting/ScriptingExecutor.java
@@ -30,4 +30,6 @@ public interface ScriptingExecutor {
 
 	public ScriptingExecutor newInstance(boolean executeInSeparateThread);
 
+	public void validate(String script) throws ScriptingException;
+
 }
